move shared service to the top level [ci:force]
