+++ lib/process-services/src/lib/app-list/select-apps-dialog/select-apps-dialog.component.spec.ts
@@ -19,7 +19,7 @@ import { Component } from '@angular/core';
 import { TestBed, ComponentFixture } from '@angular/core/testing';
 import { MatDialog, MatDialogRef, MAT_DIALOG_DATA } from '@angular/material/dialog';
 import { OverlayContainer } from '@angular/cdk/overlay';
-import { AppsProcessService } from '../services/apps-process.service';
+import { AppsProcessService } from '../../services/apps-process.service';
 import { deployedApps } from '../../mock/apps-list.mock';
 import { of } from 'rxjs';
 import { SelectAppsDialogComponent } from './select-apps-dialog.component';
