+++ lib/process-services/src/lib/app-list/select-apps-dialog/select-apps-dialog.component.ts
@@ -17,7 +17,7 @@
 
 import { Component, Inject, ViewEncapsulation } from '@angular/core';
 import { MAT_DIALOG_DATA, MatDialogModule, MatDialogRef } from '@angular/material/dialog';
-import { AppsProcessService } from '../services/apps-process.service';
+import { AppsProcessService } from '../../services/apps-process.service';
 import { CommonModule } from '@angular/common';
 import { TranslateModule } from '@ngx-translate/core';
 import { MatSelectModule } from '@angular/material/select';
